% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_costs.R
\name{adjust_costs}
\alias{adjust_costs}
\title{Adjust YJ costs to a given financial year using CPI}
\usage{
adjust_costs(
  fy = NULL,
  location = c("Australia", "Sydney", "Melbourne", "Brisbane", "Adelaide", "Perth",
    "Hobart", "Darwin", "Canberra"),
  ...
)
}
\arguments{
\item{fy}{\code{character} denoting the financial year to adjust to. Must be of the format \code{"YYYY-YY"}. Defaults to \code{NULL}}

\item{location}{\code{character} denoting which city to pull CPI data for. Can be one of \code{"Australia"}, \code{"Sydney"}, \code{"Melbourne"}, \code{"Brisbane"}, \code{"Adelaide"}, \code{"Perth"}, \code{"Hobart"}, \code{"Darwin"}, \code{"Canberra"}. Defaults to \code{"Australia"}}

\item{...}{arguments to be passed to \code{avg_cpi}}
}
\value{
\code{data.frame} of adjusted costs and associated information
}
\description{
Adjust YJ costs to a given financial year using CPI
}
\author{
Trent Henderson
}
