% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_cpi.R
\name{avg_cpi}
\alias{avg_cpi}
\title{Calculate an average CPI over a given FY}
\usage{
avg_cpi(cpi, fy = NULL, ...)
}
\arguments{
\item{cpi}{\code{cpi} object containing the inflation rates as returned by \code{get_cpi}}

\item{fy}{\code{character} denoting the financial year to calculate the average over. Must be of the format \code{"YYYY-YY"}. Defaults to \code{NULL}}

\item{...}{arguments to be passed to \code{mean}}
}
\value{
\code{numeric} scalar of the average inflation over the FY
}
\description{
Calculate an average CPI over a given FY
}
\author{
Trent Henderson
}
