% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_to_fy.R
\name{calendar_to_fy}
\alias{calendar_to_fy}
\title{Convert calendar dates to financial years}
\usage{
calendar_to_fy(x, format = NULL)
}
\arguments{
\item{x}{\code{vector} of dates}

\item{format}{\code{character} denoting the date format}
}
\value{
\code{vector} of FYs
}
\description{
Convert calendar dates to financial years
}
\examples{
calendar_to_fy("01/01/97", format = "\%d/\%m/\%Y")

}
\author{
Trent Henderson
}
