% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpi.R
\name{get_cpi}
\alias{get_cpi}
\title{Pull historical CPI data from the Australian Bureau of Statistics}
\usage{
get_cpi(
  location = c("Australia", "Sydney", "Melbourne", "Brisbane", "Adelaide", "Perth",
    "Hobart", "Darwin", "Canberra")
)
}
\arguments{
\item{location}{\code{character} denoting which city to pull CPI data for. Can be one of \code{"Australia"}, \code{"Sydney"}, \code{"Melbourne"}, \code{"Brisbane"}, \code{"Adelaide"}, \code{"Perth"}, \code{"Hobart"}, \code{"Darwin"}, \code{"Canberra"}. Defaults to \code{"Australia"}}
}
\value{
\code{data.frame} of results
}
\description{
Pull historical CPI data from the Australian Bureau of Statistics
}
\author{
Trent Henderson
}
