% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qldyjcost.R
\docType{package}
\name{qldyjcost}
\alias{qldyjcost}
\alias{qldyjcost-package}
\title{Cost Of Offending Tools For Youth Justice In Queensland}
\description{
Cost Of Offending Tools For Youth Justice In Queensland
}
